requirements.txt
pika==1.3.2
kafka-python==2.1.5
six==1.17.0
python-dotenv==1.1.0
pip==25.0.1
pika-stubs==0.1.3
isort==6.0.1
black==25.1.0